fix: update import paths after PostHogClientProvider relocation
