+++ src/services/error/providers/PostHogErrorProvider.ts
@@ -2,7 +2,7 @@ import { PostHog } from "posthog-node"
 import * as vscode from "vscode"
 import { HostProvider } from "@/hosts/host-provider"
 import { getDistinctId } from "@/services/logging/distinctId"
-import { PostHogClientProvider } from "@/services/posthog/PostHogClientProvider"
+import { PostHogClientProvider } from "@/services/telemetry/providers/PostHogClientProvider"
 import { Setting } from "@/shared/proto/index.host"
 import * as pkg from "../../../../package.json"
 import { PostHogClientValidConfig } from "../../../shared/services/config/posthog-config"
